Update inter.c
Hey Felix, I was looking at this to test out `inter.c` during the examshell and it seems the output in the end is not correctly. The reason is because the ASCII is being ordered by the lowest value of the table.

Input: "zeid"
output "ediz"

I made a quick solution for this problem is to only check it once. My first iteration with a single counter will compare the values with the parsed string and flag it if the condition it right. I hope I see you one of these days anywhere :)